api/requirements.txt
fastapi
pydantic
uvicorn
beautifulsoup4
lxml
numpy
Pillow
seaborn
python-multipart
networkx
scipy
torch
torch_geometric
psycopg2-binary